Refact : 커스텀 훅, 유틸 분리
