+++ finda2.0/src/hooks/useMove.ts
@@ -1,0 +1,16 @@
+import { useNavigate } from 'react-router-dom';
+
+interface gotoPageProps {
+  url?: 'result' | 'movie';
+  detail: string;
+}
+
+export const useMove = () => {
+  const navigate = useNavigate();
+
+  const gotoPage = ({ url, detail }: gotoPageProps) => {
+    navigate(`/${url}/${detail}`);
+  };
+
+  return gotoPage;
+};
