+++ finda2.0/src/utils/util.ts
@@ -1,7 +1,3 @@
-import { useNavigate } from 'react-router-dom';
-
-const navigate = useNavigate();
-
 interface debouncingType {
   callback: (e: React.ChangeEvent<HTMLInputElement>) => void;
   time?: number;
@@ -13,13 +9,4 @@ export const debouncing = ({ callback, time = 500 }: debouncingType) => {
     if (timerId) clearTimeout(timerId);
     timerId = setTimeout(() => callback(e), time);
   };
-};
-
-interface gotoPageProps {
-  url?: 'result' | 'movie';
-  detail: string;
-}
-
-export const gotoPage = ({ url, detail }: gotoPageProps) => {
-  navigate(`/${url}/${detail}`);
 };
